client: Initialize using a config string
The libmemcached APIs to add weighted servers to a client are not widely
available in all releases of libmemcached.

To work around this, we're now using the `memcached(const char*,
size_t)` API, which is recommended by the authors as the standard way to
initialize the client.

See: http://docs.libmemcached.org/libmemcached_configuration.html

The constructor for Memcached::Connection now takes a configuration
string. To keep the external API unchanged, Memcached::Client#new now
parses the array of servers and creates a configuration string that will
be passed to the connection.

Alternatively, you can pass directly a configuration string to
Client#new and that will be used in place of the array of servers.

A new static method Connection#check_config! has been added so we can
eagerly verify that user supplied config strings are valid.